refactor(wevu-compiler): split long source files into focused modules
