+++ packages/wevu-compiler/src/plugins/vue/compiler/template/elements/tag-slot.ts
@@ -1,8 +1,6 @@
 import type { DirectiveNode, ElementNode } from '@vue/compiler-core'
 import type { TransformContext, TransformNode } from '../types'
-import * as t from '@babel/types'
 import { NodeTypes } from '@vue/compiler-core'
-import { parse as babelParse } from '../../../../../utils/babel'
 import { buildClassStyleWxsTag } from '../classStyleRuntime'
 import { normalizeWxmlExpressionWithContext } from '../expression'
 import { renderMustache } from '../mustache'
@@ -10,9 +8,9 @@ import {
   collectScopePropMapping,
   hashString,
   isScopedSlotsDisabled,
-  toWxmlStringLiteral,
   withSlotProps,
 } from './helpers'
+import { collectSlotBindingExpression, parseSlotPropsExpression } from './slotProps'
 
 export type SlotNameInfo = { type: 'default' } | { type: 'static', value: string } | { type: 'dynamic', exp: string }
 
@@ -89,69 +87,6 @@ export function stringifySlotName(info: SlotNameInfo, context: TransformContext)
   }
   const normalized = normalizeWxmlExpressionWithContext(info.exp, context)
   return normalized
-}
-
-function parseSlotPropsExpression(exp: string, context: TransformContext): Record<string, string> {
-  const trimmed = exp.trim()
-  if (!trimmed) {
-    return {}
-  }
-  try {
-    const ast = babelParse(`(${trimmed}) => {}`, { sourceType: 'module', plugins: ['typescript'] })
-    const stmt = ast.program.body[0]
-    if (!stmt || !('expression' in stmt)) {
-      return {}
-    }
-    const expression = (stmt as any).expression as t.Expression
-    if (!t.isArrowFunctionExpression(expression)) {
-      return {}
-    }
-    const param = expression.params[0]
-    if (!param) {
-      return {}
-    }
-    if (t.isIdentifier(param)) {
-      return { [param.name]: '' }
-    }
-    if (t.isObjectPattern(param)) {
-      const mapping: Record<string, string> = {}
-      for (const prop of param.properties) {
-        if (t.isRestElement(prop)) {
-          context.warnings.push('小程序不支持作用域插槽的剩余解构元素。')
-          continue
-        }
-        if (!t.isObjectProperty(prop)) {
-          continue
-        }
-        const key = prop.key
-        const propName = t.isIdentifier(key)
-          ? key.name
-          : t.isStringLiteral(key)
-            ? key.value
-            : undefined
-        if (!propName) {
-          context.warnings.push('小程序不支持作用域插槽的计算属性键。')
-          continue
-        }
-        const value = prop.value
-        if (t.isIdentifier(value)) {
-          mapping[value.name] = propName
-          continue
-        }
-        if (t.isAssignmentPattern(value) && t.isIdentifier(value.left)) {
-          mapping[value.left.name] = propName
-          context.warnings.push('不支持作用域插槽参数的默认值，默认值将被忽略。')
-          continue
-        }
-        context.warnings.push('作用域插槽解构仅支持标识符绑定。')
-      }
-      return mapping
-    }
-  }
-  catch {
-    context.warnings.push('作用域插槽参数解析失败，已回退为空参数。')
-  }
-  return {}
 }
 
 export function buildSlotDeclaration(
@@ -235,46 +170,7 @@ export function transformSlotElement(node: ElementNode, context: TransformContex
     return transformSlotElementPlain(node, context, transformNode)
   }
   const slotNameInfo = resolveSlotNameFromSlotElement(node)
-  let bindObjectExp: string | null = null
-  const namedBindings: Array<{ key: string, value: string }> = []
-
-  for (const prop of node.props) {
-    if (prop.type === NodeTypes.ATTRIBUTE && prop.name === 'name') {
-      continue
-    }
-    if (prop.type === NodeTypes.DIRECTIVE && prop.name === 'bind') {
-      if (prop.arg?.type === NodeTypes.SIMPLE_EXPRESSION) {
-        const rawExpValue = prop.exp?.type === NodeTypes.SIMPLE_EXPRESSION ? prop.exp.content : ''
-        if (prop.arg.content === 'name') {
-          continue
-        }
-        if (rawExpValue) {
-          namedBindings.push({ key: prop.arg.content, value: normalizeWxmlExpressionWithContext(rawExpValue, context) })
-        }
-        continue
-      }
-      if (prop.exp?.type === NodeTypes.SIMPLE_EXPRESSION) {
-        bindObjectExp = normalizeWxmlExpressionWithContext(prop.exp.content, context)
-        continue
-      }
-    }
-    if (prop.type === NodeTypes.ATTRIBUTE && prop.name !== 'name') {
-      const literal = prop.value?.type === NodeTypes.TEXT ? prop.value.content : ''
-      if (literal) {
-        namedBindings.push({ key: prop.name, value: `'${literal.replace(/\\/g, '\\\\').replace(/'/g, '\\\'')}'` })
-      }
-    }
-  }
-
-  if (bindObjectExp && namedBindings.length) {
-    context.warnings.push('作用域插槽参数使用 v-bind 对象时，将忽略额外的命名绑定。')
-    namedBindings.length = 0
-  }
-
-  let slotPropsExp = bindObjectExp
-  if (!slotPropsExp && namedBindings.length) {
-    slotPropsExp = `[${namedBindings.map(entry => `${toWxmlStringLiteral(entry.key)},${entry.value}`).join(',')}]`
-  }
+  let slotPropsExp = collectSlotBindingExpression(node, context)
 
   let fallbackContent = ''
   if (node.children.length > 0) {
@@ -306,10 +202,10 @@ export function transformSlotElement(node: ElementNode, context: TransformContex
   const genericKey = `scoped-slots-${slotKey}`
   context.componentGenerics[genericKey] = true
 
-  const resolvedSlotPropsExp = slotPropsExp ?? '[]'
+  slotPropsExp = slotPropsExp ?? '[]'
   const scopedAttrs = [
     `__wv-owner-id="${renderMustache('__wvSlotOwnerId', context)}"`,
-    `__wv-slot-props="${renderMustache(resolvedSlotPropsExp, context)}"`,
+    `__wv-slot-props="${renderMustache(slotPropsExp, context)}"`,
   ]
   if (context.slotMultipleInstance) {
     scopedAttrs.push(`__wv-slot-scope="${renderMustache('__wvSlotScope', context)}"`)
